+++ src/lib/compute/GPU.cpp
@@ -4,7 +4,7 @@
  * Created:
  *   16/04/2021, 17:21:49
  * Last edited:
- *   23/05/2021, 21:17:19
+ *   25/05/2021, 17:23:26
  * Auto updated?
  *   Yes
  *
@@ -14,7 +14,7 @@
  *   library.
 **/
 
-#include <CppDebugger.hpp>
+#include "debugger/CppDebugger.hpp"
 
 #include "compute/ErrorCodes.hpp"
 
@@ -255,7 +255,7 @@ void DeviceQueueInfo::check_present(VkPhysicalDevice vk_physical_device, VkSurfa
     for (size_t i = 0; i < supported_queues.size(); i++) {
         // Query the support of this queue
         VkBool32 supports_presenting;
-        vkGetPhysicalDeviceSurfaceSupportKHR(vk_physical_device, i, vk_surface, &supports_presenting);
+        vkGetPhysicalDeviceSurfaceSupportKHR(vk_physical_device, (uint32_t) i, vk_surface, &supports_presenting);
 
         // If it supports presenting, then set it if a) it's the first or b) it's different from any of the other found queues
         if (supports_presenting == VK_TRUE && (!this->supports_presentation || (this->supports_compute && this->presentation_index != this->compute_index && this->supports_memory && this->presentation_index != this->memory_index))) {
